They should be called SU Registers (Setup Unit/Rasterization).  BP (Bypass) is really the name of the commands that are passed in.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@3531 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Plugins/Plugin_VideoOGL/Src/rasterfont.cpp
@@ -141,7 +141,7 @@ RasterFont::~RasterFont()
 
 void RasterFont::printString(const char *s, double x, double y, double z)
 {
-	int length = strlen(s);
+	int length = (int)strlen(s);
 	if (!length)
 		return;
